+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Missing e-mail prefix in contact form
 - Driver Age function when birthdate is None
+- Accidentally deleted get_initial in the contact form
 
 ## [2.7.8] - 2024-10-29
 ### Changed